BGDIINF_SB-1711: Moved paginator in its own module
