+++ app/config/settings_prod.py
@@ -270,7 +270,7 @@ REST_FRAMEWORK = {
         'rest_framework.authentication.TokenAuthentication',
         'rest_framework.authentication.SessionAuthentication',
     ],
-    'DEFAULT_PAGINATION_CLASS': 'stac_api.apps.CursorPagination',
+    'DEFAULT_PAGINATION_CLASS': 'stac_api.pagination.CursorPagination',
     'PAGE_SIZE': os.environ.get('PAGE_SIZE', 100),
     'PAGE_SIZE_LIMIT': os.environ.get('PAGE_SIZE_LIMIT', 100),
     'EXCEPTION_HANDLER': 'stac_api.apps.custom_exception_handler',
